+++ packages/providers/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hookfish/provider-github
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [397067d]
+  - @hookfish/provider@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes